+++ erl/timer.go
@@ -6,12 +6,26 @@ import (
 	"github.com/uberbrodt/erl-go/erl/exitreason"
 )
 
+type TimerRef struct {
+	pid PID
+}
+
+func CancelTimer(tr TimerRef) error {
+	if tr.pid.IsNil() {
+		return exitreason.NoProc
+	}
+	Send(tr.pid, cancelTimer{})
+	return nil
+}
+
 type timer struct {
 	to   PID
 	term any
 	tout time.Duration
 	ref  Ref
 }
+
+type cancelTimer struct{}
 
 func (t *timer) Receive(self PID, inbox <-chan any) error {
 	t.ref = Monitor(self, t.to)
@@ -23,6 +37,8 @@ func (t *timer) Receive(self PID, inbox <-chan any) error {
 				return exitreason.Normal
 			}
 			switch msg := anyMsg.(type) {
+			case cancelTimer:
+				return exitreason.Normal
 			case DownMsg:
 				// if the process we want to send a message to dies, cancel the timer by exiting
 				if t.ref == msg.Ref {